Add files via upload
This is a simple python script that pulls system information for Lazy and/or frustrated administrators. 
This will get you the Machine name, the IP address, and the username of the user currently logged in. No guess, no hand holding, and no heads beaten against desks.
I've included the compiled exe along with the code which I hope to add on to as time allows.